src/test/java/com/github/brunomndantas/jscrapper/support/parser/array/attribute/primitive/ArrayPrimitiveByteAttributeParserTest.java
package com.github.brunomndantas.jscrapper.support.parser.array.attribute.primitive;

import com.github.brunomndantas.jscrapper.DummyElement;
import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertSame;

public class ArrayPrimitiveByteAttributeParserTest {

    @Test
    public void constructorTest() {
        String attribute = "";
        ArrayPrimitiveByteAttributeParser parser = new ArrayPrimitiveByteAttributeParser(attribute);
        assertSame(byte.class, parser.getKlass());
        assertEquals((byte)0, parser.getDefaultValue());
        assertSame(attribute, parser.getAttribute());
    }

    @Test
    public void parseElement() throws Exception {
        ArrayPrimitiveByteAttributeParser parser = new ArrayPrimitiveByteAttributeParser("");

        assertEquals((byte)2, parser.parseElement(new DummyElement(){
            @Override
            public String getAttribute(String a) {
                return "2";
            }
        }));
    }

}
